+++ components/Chat/Message/MessageReactions/MessageReactions.tsx
@@ -3,13 +3,16 @@ import { HStack } from "@design-system/HStack";
 import { Text } from "@design-system/Text";
 import { VStack } from "@design-system/VStack";
 import { useAppTheme } from "@theme/useAppTheme";
-import { memo, useCallback } from "react";
+import { memo, useCallback, useMemo } from "react";
 import { StyleSheet, TouchableHighlight } from "react-native";
 
 import { MessageToDisplay } from "../Message";
-import { MessageReactions } from "./MessageReactions.types";
-import { openMessageReactionsDrawer } from "./MessageReactionsDrawer/MessageReactionsDrawer.utils";
-import { useMessageReactionsRolledup } from "./useMessageReactionsRolledup";
+import {
+  MessageReactions,
+  ReactionDetails,
+  RolledUpReactions,
+} from "./MessageReactions.types";
+import { openMessageReactionsDrawer } from "./MessageReactionsDrawer/MessageReactionsDrawer.service";
 
 type Props = {
   message: MessageToDisplay;
@@ -28,8 +31,8 @@ export const ChatMessageReactions = memo(
     });
 
     const handlePress = useCallback(() => {
-      openMessageReactionsDrawer(message);
-    }, [message]);
+      openMessageReactionsDrawer(rolledUpReactions);
+    }, [rolledUpReactions]);
 
     if (rolledUpReactions.totalReactions === 0) return null;
 
@@ -104,3 +107,54 @@ const useStyles = () => {
     },
   });
 };
+
+const MAX_REACTION_EMOJIS_SHOWN = 3;
+
+function useMessageReactionsRolledup(arg: {
+  reactions: MessageReactions;
+  userAddress: string;
+}) {
+  const { reactions, userAddress } = arg;
+
+  return useMemo((): RolledUpReactions => {
+    const details: { [content: string]: ReactionDetails } = {};
+    let totalReactions = 0;
+    let userReacted = false;
+
+    Object.values(reactions).forEach((reactionArray) => {
+      reactionArray.forEach((reaction) => {
+        if (!details[reaction.content]) {
+          details[reaction.content] = {
+            content: reaction.content,
+            count: 0,
+            userReacted: false,
+            reactors: [],
+            firstReactionTime: reaction.sent,
+          };
+        }
+        details[reaction.content].count++;
+        details[reaction.content].reactors.push(reaction.senderAddress);
+        if (
+          reaction.senderAddress.toLowerCase() === userAddress?.toLowerCase()
+        ) {
+          details[reaction.content].userReacted = true;
+          userReacted = true;
+        }
+        // Keep track of the earliest reaction time for this emoji
+        details[reaction.content].firstReactionTime = Math.min(
+          details[reaction.content].firstReactionTime,
+          reaction.sent
+        );
+        totalReactions++;
+      });
+    });
+
+    // Sort by the number of reactors in descending order
+    const sortedReactions = Object.values(details)
+      .sort((a, b) => b.reactors.length - a.reactors.length)
+      .slice(0, MAX_REACTION_EMOJIS_SHOWN)
+      .map((reaction) => reaction.content);
+
+    return { emojis: sortedReactions, totalReactions, userReacted, details };
+  }, [reactions, userAddress]);
+}
